Tagger GUI now using VideoFile class. Can take subsection of video after 1 more small change to entry.py
