+++ src/gui_tool/GUIExceptions.py
@@ -4,9 +4,3 @@ class ManualTaggingAbortedException(Exception):
 
 class ManualTaggingExitedException(Exception):
     '''Raise when user aborts the tagging'''
-
-class VideoNotFoundException(Exception):
-    '''Raise when user aborts the tagging'''
-
-class VideoCouldNotBeOpenedException(Exception):
-    '''Raise when user aborts the tagging'''
