+++ src/gui_tool/gui/context.py
@@ -2,9 +2,9 @@
 from ..VideoCaptureManager import VideoCaptureManager
 
 class GUIContext(object):
-    def __init__(self, file_loc: str):
+    def __init__(self, file_loc: str, start_index=None, end_index=None):
         self.file_loc = file_loc
-        self.vcm = VideoCaptureManager(file_loc)
+        self.vcm = VideoCaptureManager(file_loc, start_index, end_index)
         self.vcm.start_from(0)
         self.file_height = self.vcm.get_height()
         self.file_width = self.vcm.get_width()
